add two IKE text display functions
there are 2 ecus which also support writing on the text display of the IKE: RAD and CCM (aka the LCM). The message of the LCM/CCM must be deleted manually (that is what happens when releasing the check control button). The message of the RAD can (contrary to all available documentation) also display 20 chars. The function fills up shorter messages with trailing 0x20 spaces.